requirements.txt
# GitHub API
PyGithub==2.5.0
requests==2.32.3

# Google Gemini AI
google-generativeai==0.8.3

# Data Processing
python-dateutil==2.9.0
pytz==2024.2

# Visualization
matplotlib==3.9.2
plotly==5.24.1
kaleido==0.2.1

# Email
python-dotenv==1.0.1

# YAML Configuration
PyYAML==6.0.2
